fix: widen CHECK_EQ to long long, drop the 21 (int) casts
The check.h harness compared everything as int, so routing the Track A/C
harnesses through it required wrapping call sites as CHECK_EQ((int)(x),
(int)(y)). That truncates the exact property under test in the exercises
that are *about* width: 04-args-regs passes longs in argument registers,
05-return-values returns a long in rax, 82 counts bytes as size_t. A
64-bit return that differed above 2^31 would have passed silently.

Compare as long long and print with %lld; remove the now-redundant casts
from all 21 sites. The failure text also stops echoing the cast noise.

+++ 04-args-regs/prog.c
@@ -37,10 +37,10 @@ long f7(long a, long b, long c, long d, long e, long f, long g)
 int main(void)
 {
     /* Sums: f1(1)=1, f2(1,2)=3, f6(1..6)=21, f7(1..7)=28 */
-    CHECK_EQ((int)(f1(1)), (int)(1));
-    CHECK_EQ((int)(f2(1, 2)), (int)(3));
-    CHECK_EQ((int)(f6(1, 2, 3, 4, 5, 6)), (int)(21));
-    CHECK_EQ((int)(f7(1, 2, 3, 4, 5, 6, 7)), (int)(28));
+    CHECK_EQ(f1(1), 1);
+    CHECK_EQ(f2(1, 2), 3);
+    CHECK_EQ(f6(1, 2, 3, 4, 5, 6), 21);
+    CHECK_EQ(f7(1, 2, 3, 4, 5, 6, 7), 28);
     if (test_report() != 0)
         return 1;
     puts("ok");
